updating the image uploading feature
