+++ src/main/java/com/example/vehiclerentingapplication/controller/UserController.java
@@ -33,12 +33,5 @@ public class UserController {
 				.body(ResponseStructure.create(HttpStatus.CREATED.value(), "User Created", user));
 	}
 	
-	  @PostMapping("/users/profile-picture")
-	    public ResponseEntity<SimpleResponseStructure> addProfilePicture(
-	            @RequestParam("userId") int userId, 
-	            @RequestParam("file") MultipartFile file) throws IOException  {
-	       userService.addUserProfilePicture(userId, file);
-	        return ResponseEntity.status(HttpStatus.OK)
-	                .body(SimpleResponseStructure.create(HttpStatus.OK.value(), "Profile picture added successfully"));
-	    }
+	 
 }
