+++ src/main/java/com/example/vehiclerentingapplication/service/ImageService.java
@@ -1,0 +1,58 @@
+package com.example.vehiclerentingapplication.service;
+
+import java.io.IOException;
+import java.util.Optional;
+
+import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
+
+import com.example.vehiclerentingapplication.entity.Image;
+import com.example.vehiclerentingapplication.entity.User;
+import com.example.vehiclerentingapplication.exception.FailedToUploadImageException;
+import com.example.vehiclerentingapplication.exception.UserNotFoundByIdException;
+import com.example.vehiclerentingapplication.repository.ImageRepository;
+import com.example.vehiclerentingapplication.repository.UserRepository;
+
+@Service
+public class ImageService {
+
+	private final ImageRepository imageRepository;
+
+	private final UserRepository userRepository;
+
+	public ImageService(ImageRepository imageRepository, UserRepository repository) {
+		super();
+		this.imageRepository = imageRepository;
+		this.userRepository = repository;
+	}
+
+	public void addUserProfilePicture(int userId, MultipartFile multipartFile) {
+
+		Optional<User> optional = userRepository.findById(userId);
+
+		if (optional.isPresent()) {
+
+			Image image = ImageService.getImage(multipartFile);
+			image = imageRepository.save(image);
+
+			User user = optional.get();
+			user.setProfilePicture(image);
+			userRepository.save(user);
+		} else {
+			throw new UserNotFoundByIdException("Failed to Find the user");
+		}
+	}
+
+	public static Image getImage(MultipartFile imagefile) {
+		Image image = new Image();
+		try {
+			byte[] imageBytes = imagefile.getBytes();
+			image.setContentType(imagefile.getContentType());
+			image.setImageBytes(imageBytes);
+		} catch (IOException e) {
+			throw new FailedToUploadImageException("Failed to upload the Image");
+		}
+
+		return image;
+	}
+}
